replicando o comportamento do aviso
